refactor : Auth 도메인 리팩토링
