+++ backend/src/main/java/com/graphy/backend/domain/auth/controller/AuthController.java
@@ -1,11 +1,12 @@
 package com.graphy.backend.domain.auth.controller;
 
+import com.graphy.backend.domain.auth.dto.request.LogoutRequest;
+import com.graphy.backend.domain.auth.dto.response.GetTokenInfoResponse;
 import com.graphy.backend.domain.auth.service.AuthService;
 import com.graphy.backend.domain.auth.util.annotation.CurrentUser;
 import com.graphy.backend.domain.member.domain.Member;
 import com.graphy.backend.domain.member.dto.request.SignInMemberRequest;
 import com.graphy.backend.domain.member.dto.request.SignUpMemberRequest;
-import com.graphy.backend.domain.auth.dto.TokenDto;
 import com.graphy.backend.global.result.ResultCode;
 import com.graphy.backend.global.result.ResultResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -21,7 +22,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
 
-import static com.graphy.backend.domain.auth.dto.TokenDto.*;
 
 @Tag(name = "AuthController", description = "인증 인가 API")
 @RestController
@@ -38,9 +38,8 @@ public class AuthController {
     
     @Operation(summary = "signin", description = "로그인")
     @PostMapping("/signin")
-    public TokenInfo signIn(HttpServletRequest request,
-                                     @RequestBody SignInMemberRequest dto) {
-        return authService.signIn(request, dto);
+    public GetTokenInfoResponse signIn(@Validated @RequestBody SignInMemberRequest dto) {
+        return authService.signIn(dto);
     }
 
     @Operation(summary = "logout", description = "로그아웃")
@@ -52,8 +51,8 @@ public class AuthController {
     
     @Operation(summary = "reIssue", description = "토큰 재발급")
     @PostMapping(value = "/reissue")
-    public TokenInfo reissue(HttpServletRequest request,
-                             @CurrentUser Member member) {
+    public GetTokenInfoResponse reissue(HttpServletRequest request,
+                                        @CurrentUser Member member) {
         return authService.reissue(request, member);
     }
 }
