--- backend/src/main/java/com/graphy/backend/domain/auth/dto/TokenDto.java
@@ -1,27 +1,0 @@
-package com.graphy.backend.domain.auth.dto;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-
-public class TokenDto {
-    @Getter
-    @Builder
-    @AllArgsConstructor(access = AccessLevel.PRIVATE)
-    public static class TokenInfo {
-        private String grantType;
-        private String accessToken;
-        private String refreshToken;
-    }
-
-    @Getter
-    @Builder
-    @AllArgsConstructor(access = AccessLevel.PRIVATE)
-    public static class LogoutRequest {
-        private String accessToken;
-        private String refreshToken;
-    }
-
-}
