+++ backend/src/main/java/com/graphy/backend/domain/auth/dto/request/LogoutRequest.java
@@ -1,0 +1,14 @@
+package com.graphy.backend.domain.auth.dto.request;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class LogoutRequest {
+    private String accessToken;
+    private String refreshToken;
+}
