+++ backend/src/main/java/com/graphy/backend/domain/auth/dto/response/GetTokenInfoResponse.java
@@ -1,0 +1,15 @@
+package com.graphy.backend.domain.auth.dto.response;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class GetTokenInfoResponse {
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+}
