+++ backend/src/main/java/com/graphy/backend/domain/auth/service/AuthService.java
@@ -1,16 +1,14 @@
 package com.graphy.backend.domain.auth.service;
 
 import com.graphy.backend.domain.auth.controller.JwtFilter;
+import com.graphy.backend.domain.auth.dto.request.LogoutRequest;
+import com.graphy.backend.domain.auth.dto.response.GetTokenInfoResponse;
 import com.graphy.backend.domain.member.domain.Member;
 import com.graphy.backend.domain.member.dto.request.SignInMemberRequest;
 import com.graphy.backend.domain.member.dto.request.SignUpMemberRequest;
-import com.graphy.backend.domain.member.repository.MemberRepository;
 import com.graphy.backend.domain.member.service.MemberService;
 import com.graphy.backend.domain.auth.domain.RefreshToken;
 import com.graphy.backend.domain.auth.repository.RefreshTokenRepository;
-import com.graphy.backend.global.common.Helper;
-import com.graphy.backend.global.error.ErrorCode;
-import com.graphy.backend.global.error.exception.AlreadyExistException;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -21,7 +19,6 @@ import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpServletRequest;
 
-import static com.graphy.backend.domain.auth.dto.TokenDto.*;
 
 @Service
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
@@ -41,8 +38,7 @@ public class AuthService {
         memberService.addMember(member);
     }
 
-    public TokenInfo signIn(HttpServletRequest request,
-                                     SignInMemberRequest dto) {
+    public GetTokenInfoResponse signIn(SignInMemberRequest dto) {
 
         UsernamePasswordAuthenticationToken authenticationToken =
                 new UsernamePasswordAuthenticationToken(dto.getEmail(), dto.getPassword());
@@ -50,13 +46,12 @@ public class AuthService {
         Authentication authentication = authenticationManagerBuilder.getObject()
                 .authenticate(authenticationToken);
 
-        TokenInfo token = tokenProvider.generateToken(authentication);
+        GetTokenInfoResponse token = tokenProvider.generateToken(authentication);
 
         RefreshToken refreshToken = RefreshToken.builder()
                 .token(token.getRefreshToken())
                 .email(dto.getEmail())
                 .authorities(authentication.getAuthorities())
-                .ip(Helper.getClientIp(request))
                 .build();
 
         refreshTokenRepository.save(refreshToken);
@@ -92,36 +87,30 @@ public class AuthService {
         refreshTokenRepository.save(token);
     }
 
-    public TokenInfo reissue(HttpServletRequest request, Member member) {
-        //1. Request Header 에서 JWT Token 추출
+    public GetTokenInfoResponse reissue(HttpServletRequest request, Member member) {
+        // Request Header 에서 JWT Token 추출
         String requestToken = filter.resolveToken(request);
 
-        //2. validateToken 메서드로 토큰 유효성 검사
+        // validateToken 메서드로 토큰 유효성 검사
         if (requestToken != null && tokenProvider.validateToken(requestToken)) {
 
-            //3. 저장된 refresh token 찾기
+            // 저장된 refresh token 찾기
             RefreshToken savedRefreshToken = refreshTokenRepository.findByEmail(member.getEmail());
             if (savedRefreshToken != null) {
 
-                //4. 최초 로그인한 ip 와 같은지 확인
-                String currentIpAddress = Helper.getClientIp(request);
-                if (savedRefreshToken.getIp().equals(currentIpAddress)) {
+                // Redis 에 저장된 RefreshToken 정보를 기반으로 JWT Token 생성
+                GetTokenInfoResponse newToken = tokenProvider.generateToken(savedRefreshToken.getToken(), savedRefreshToken.getAuthorities());
 
-                    // 5. Redis 에 저장된 RefreshToken 정보를 기반으로 JWT Token 생성
-                    TokenInfo newToken = tokenProvider.generateToken(savedRefreshToken.getToken(), savedRefreshToken.getAuthorities());
+                // Redis RefreshToken update
+                RefreshToken newRefreshToken = RefreshToken.builder()
+                        .token(newToken.getRefreshToken())
+                        .email(member.getEmail())
+                        .authorities(savedRefreshToken.getAuthorities())
+                        .build();
 
-                    // 6. Redis RefreshToken update
-                    RefreshToken newRefreshToken = RefreshToken.builder()
-                            .token(newToken.getRefreshToken())
-                            .email(member.getEmail())
-                            .ip(savedRefreshToken.getIp())
-                            .authorities(savedRefreshToken.getAuthorities())
-                            .build();
-
-                    refreshTokenRepository.save(newRefreshToken);
-                    savedRefreshToken.updateToken(newRefreshToken.getToken());
-                    return newToken;
-                }
+                refreshTokenRepository.save(newRefreshToken);
+                savedRefreshToken.updateToken(newRefreshToken.getToken());
+                return newToken;
             }
         }
         return null;
