--- backend/src/main/java/com/graphy/backend/global/common/Helper.java
@@ -1,44 +1,0 @@
-package com.graphy.backend.global.common;
-
-import org.springframework.util.StringUtils;
-import javax.servlet.http.HttpServletRequest;
-import java.util.ArrayList;
-import java.util.List;
-
-public class Helper {
-
-    // 유틸리티 클래스의 불필요한 인스턴스화를 방지
-    private Helper() {
-    }
-
-    public static String getClientIp(HttpServletRequest request) {
-        String clientIp = null;
-        boolean isIpInHeader = false;
-
-        List<String> headerList = new ArrayList<>();
-        headerList.add("X-Forwarded-For");
-        headerList.add("HTTP_CLIENT_IP");
-        headerList.add("HTTP_X_FORWARDED_FOR");
-        headerList.add("HTTP_X_FORWARDED");
-        headerList.add("HTTP_FORWARDED_FOR");
-        headerList.add("HTTP_FORWARDED");
-        headerList.add("Proxy-Client-IP");
-        headerList.add("WL-Proxy-Client-IP");
-        headerList.add("HTTP_VIA");
-        headerList.add("IPV6_ADR");
-
-        for (String header : headerList) {
-            clientIp = request.getHeader(header);
-            if (StringUtils.hasText(clientIp) && !clientIp.equals("unknown")) {
-                isIpInHeader = true;
-                break;
-            }
-        }
-
-        if (!isIpInHeader) {
-            clientIp = request.getRemoteAddr();
-        }
-
-        return clientIp;
-    }
-}
